+++ internal/operations/read_balance.go
@@ -4,6 +4,6 @@ import "github.com/rwwae/simplebank/internal/accounts"
 
 //RetrieveBalance is pass through operation that returns the accountBalance unchanged.
 var RetrieveBalance = accounts.BalanceOperationFunc(
-	func(accountBalance float64) (float64, error) {
+	func(accountBalance int64) (int64, error) {
 		return accountBalance, nil
 	})
